backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Cosecre FastAPI backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "openai>=2.30.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.43.0",
]

[project.scripts]
backend = "backend.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
